+++ pkg/apiclient/http/facade_test.go
@@ -9,8 +9,8 @@ import (
 
 func TestFacade_do(t *testing.T) {
 	f := Facade{baseUrl: "http://my-url"}
-	u, err := f.url("GET", "/{name}", &metav1.ObjectMeta{Name: "my-name", Labels: map[string]string{"foo": "1"}})
+	u, err := f.url("GET", "/{namespace}/{name}", &metav1.ObjectMeta{Namespace: "my-ns", Labels: map[string]string{"foo": "1"}})
 	if assert.NoError(t, err) {
-		assert.Equal(t, "http://my-url/my-name?labels.foo=1", u.String())
+		assert.Equal(t, "http://my-url/my-ns/?labels.foo=1", u.String())
 	}
 }
