+++ test/e2e/fixtures/e2e_suite.go
@@ -2,6 +2,7 @@ package fixtures
 
 import (
 	"encoding/base64"
+	"os"
 	"strings"
 	"time"
 
@@ -119,7 +120,17 @@ func (s *E2ESuite) DeleteResources() {
 	}
 }
 
-func (s *E2ESuite) AfterTest(_, _ string) {}
+func (s *E2ESuite) NeedsCI() {
+	if os.Getenv("CI") != "true" {
+		s.T().Skip("test needs CI")
+	}
+}
+
+func (s *E2ESuite) NeedsOffloading() {
+	if !s.Persistence.IsEnabled() {
+		s.T().Skip("test needs offloading, but persistence not enabled")
+	}
+}
 
 func (s *E2ESuite) dynamicFor(r schema.GroupVersionResource) dynamic.ResourceInterface {
 	resourceInterface := dynamic.NewForConfigOrDie(s.RestConfig).Resource(r)
